split frontend/parser.rs from frontend/ast.rs and frontend/translate.rs from frontend.rs
